Merged PR 7: Refactor services
Services are refactored so they enable getting sorted list of elements, all elements or within a specific range.
The main template for service interfaces is generalized into the CRUD service template to eliminate the need for NotImplementedException and then specialized with specific methods like GetAll and GetRange.
I renamed some methods, and replace names for GetSingleOrDefault and GetOne. Also, I changed the parameter for GetSingleOrDefault to move conversion logic to ServiceExecutor, to be consistent.